doc: extract template implementations from matx.hpp
